chore: extract ui methods into ui service
